Backtesting on 2017 data showed a success.
Trick is to set a threshold on Q-values when choosing an
action. Qthresh=1.25 seems to work the best
(never tried the values higher than this, but the profit scheme
and the winning rate seems to be solid for this value).
Please run the backtest.py and see it with your own eyes.

The beta version of the project is thus compete.